+++ packages/pi-gpt-5/src/index.ts
@@ -1,1 +1,3 @@
 export * from "./features.js";
+export * from "./core-pack.js";
+export * from "./entitlement.js";